% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChattelMortgage.R
\name{ChattelMortgageUI}
\alias{ChattelMortgageUI}
\title{动产抵押生成生成器界面}
\usage{
ChattelMortgageUI(
  tabTitle = "动产抵押",
  colTitles = c("操作区域", "操作区域", "显示区域"),
  widthRates = c(6, 6, 12),
  func_left = ChattelMortgageUI_left,
  func_right = ChattelMortgageUI_right,
  func_bottom = ChattelMortgageUI_bottom
)
}
\arguments{
\item{tabTitle}{标题}

\item{colTitles}{主页标题}

\item{widthRates}{左右比例}

\item{func_left}{左函数}

\item{func_right}{右函数}

\item{func_bottom}{下面一栏}
}
\value{
返回值
}
\description{
动产抵押生成生成器界面
}
\examples{
ChattelMortgageUI()
}
