% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InvestmentList.R
\name{InvestmentListUI}
\alias{InvestmentListUI}
\title{对外投资生成生成器界面}
\usage{
InvestmentListUI(
  tabTitle = "对外投资",
  colTitles = c("操作区域", "操作区域", "显示区域"),
  widthRates = c(6, 6, 12),
  func_left = InvestmentListUI_left,
  func_right = InvestmentListUI_right,
  func_bottom = InvestmentListUI_bottom
)
}
\arguments{
\item{tabTitle}{标题}

\item{colTitles}{主页标题}

\item{widthRates}{左右比例}

\item{func_left}{左函数}

\item{func_right}{右函数}

\item{func_bottom}{下面一栏}
}
\value{
返回值
}
\description{
对外投资生成生成器界面
}
\examples{
InvestmentListUI()
}
