% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShiXin.R
\name{ShiXinUI}
\alias{ShiXinUI}
\title{失信被执行人生成生成器界面}
\usage{
ShiXinUI(
  tabTitle = "失信被执行人",
  colTitles = c("操作区域", "操作区域", "显示区域"),
  widthRates = c(6, 6, 12),
  func_left = ShiXinUI_left,
  func_right = ShiXinUI_right,
  func_bottom = ShiXinUI_bottom
)
}
\arguments{
\item{tabTitle}{标题}

\item{colTitles}{主页标题}

\item{widthRates}{左右比例}

\item{func_left}{左函数}

\item{func_right}{右函数}

\item{func_bottom}{下面一栏}
}
\value{
返回值
}
\description{
失信被执行人生成生成器界面
}
\examples{
ShiXinUI()
}
