% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxCreditList.R
\name{TaxCreditListUI}
\alias{TaxCreditListUI}
\title{纳税信用等级生成生成器界面}
\usage{
TaxCreditListUI(
  tabTitle = "纳税信用等级",
  colTitles = c("操作区域", "操作区域", "显示区域"),
  widthRates = c(6, 6, 12),
  func_left = TaxCreditListUI_left,
  func_right = TaxCreditListUI_right,
  func_bottom = TaxCreditListUI_bottom
)
}
\arguments{
\item{tabTitle}{标题}

\item{colTitles}{主页标题}

\item{widthRates}{左右比例}

\item{func_left}{左函数}

\item{func_right}{右函数}

\item{func_bottom}{下面一栏}
}
\value{
返回值
}
\description{
纳税信用等级生成生成器界面
}
\examples{
TaxCreditListUI()
}
